Update SingleProductPage tests
